+++ docs/content/misc/release_notes.rst
@@ -11,6 +11,7 @@ Release notes
     - us4R: exposed SYNC mode
     - Implemented File device (for simulated mode) #221, #ARRUS-248.
     - us4R: exposed us4OEM and digital backplane serial number and revision #ARRUS-267.
+    - Added Session::getCurrentState method, which returns the current state of the session (e.g if the devices are running or not).
     
 - Python API: